added feature token authentication, use token to modify data, + uiBuild

+++ controllers/promotions.js
@@ -1,33 +1,40 @@
-const promoRouter = require('express').Router()
-const Promotion = require('../models/promotion')
+const promoRouter = require("express").Router();
+const Promotion = require("../models/promotion");
+const tokenDecoder = require("../utils/tokenDecoder");
+require("dotenv").config();
 
 //get data
-promoRouter.get('/', async (req, res) => {
-    try {
-        const result = await Promotion.findOne({})
-        res.status(200).json(result)
-    } catch (error) {
-        return res.status(400).json({ error: error.message })
-    }
-})
+promoRouter.get("/", async (req, res) => {
+  try {
+    const result = await Promotion.findOne({});
+    res.status(200).json(result);
+  } catch (error) {
+    return res.status(400).json({ error: error.message });
+  }
+});
 
 //add new obj
-promoRouter.post('/', async (req, res) => {
-    const { onPromotion, text } = req.body
+promoRouter.post("/", async (req, res) => {
+  const verifiedWithToken = tokenDecoder(req);
+  if (!verifiedWithToken) {
+    return res.status(401).json({ error: "failed to authenticate with token" });
+  }
 
-    if (onPromotion === undefined || text === undefined) {
-        return res.status(400).json({error: 'missing credentials'})
-    }
+  const { onPromotion, text } = req.body;
 
-    try {
-        await Promotion.deleteMany({})
-        const createdPromoObj = new Promotion({ onPromotion, text })
+  if (onPromotion === undefined || text === undefined) {
+    return res.status(400).json({ error: "missing credentials" });
+  }
 
-        const result = await createdPromoObj.save()
-        res.status(201).json(result)
-    } catch (error) {
-        res.status(400).json({error: error})
-    }
-})
+  try {
+    await Promotion.deleteMany({});
+    const createdPromoObj = new Promotion({ onPromotion, text });
 
-module.exports = promoRouter
+    const result = await createdPromoObj.save();
+    res.status(201).json(result);
+  } catch (error) {
+    res.status(400).json({ error: error });
+  }
+});
+
+module.exports = promoRouter;
